telegram_bot/requirements.txt
python-telegram-bot>=21.6
openai>=1.40.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.4.0
